# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TM Seasons API Docs",
        "version": "1.0.0",
        "description": "REST API for TM Seasons, covering Discord authentication, Trackmania club and map administration, season lifecycle management, leaderboard polling, event-based scoring, standings, and activity feeds."
    },
    "servers": [
        {
            "url": "https://tm-bot.test/api",
            "description": "Local"
        }
    ],
...
    "security": [
        {
            "sanctum": []
        }
    ],
    "paths": {
...
        "/seasons/{slug}/leaderboard": {
            "get": {
                "operationId": "seasons.leaderboard",
                "description": "Returns each active season map with ordered player records.",
                "summary": "Get public season leaderboard",
                "tags": [
                    "Seasons - Public"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "season": {
                                                    "$ref": "#/components/schemas/SeasonResource"
                                                },
                                                "leaderboard": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "season",
                                                "leaderboard"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/seasons/{slug}/maps/{map}/leaderboard": {
            "get": {
                "operationId": "seasons.maps.leaderboard",
                "description": "Returns ordered player records for one map within a season.",
                "summary": "Get public map leaderboard",
                "tags": [
                    "Seasons - Public"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "map",
                        "in": "path",
                        "required": true,
                        "description": "The map ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "map": {
                                                    "$ref": "#/components/schemas/MapResource"
                                                },
                                                "entries": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/SeasonMapPlayerRecordResource"
                                                    }
                                                }
                                            },
                                            "required": [
                                                "map",
                                                "entries"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                },
                "security": []
            }
        },
        "/seasons/{slug}/standings": {
            "get": {
                "operationId": "seasons.standings",
                "description": "Returns calculated point standings for a season by slug.",
                "summary": "Get public season standings",
                "tags": [
                    "Seasons - Public"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "season_id": {
                                                    "type": "integer"
                                                },
                                                "season_name": {
                                                    "type": "string"
                                                },
                                                "standings": {
                                                    "type": "object"
                                                }
                                            },
                                            "required": [
                                                "season_id",
                                                "season_name",
                                                "standings"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/admin/seasons/{season}/records": {
            "get": {
                "operationId": "admin.seasons.records",
                "description": "Returns paginated player records for a season, optionally filtered by map or player.",
                "summary": "List season map records",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "map_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `SeasonMapPlayerRecordResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SeasonMapPlayerRecordResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/seasons/{season}/recalculate": {
            "post": {
                "operationId": "admin.seasons.recalculate",
                "description": "Rebuilds point events for a season and returns the updated event count.",
                "summary": "Recalculate season points",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "Recalculation complete."
                                        },
                                        "events_count": {
                                            "type": "integer",
                                            "minimum": 0
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "events_count"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Laravel Sanctum bearer token for protected endpoints.",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "MapResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uid": {
                        "type": "string"
                    },
                    "nadeo_map_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "author_account_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "author_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "author_time": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "gold_time": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "silver_time": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "bronze_time": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "map_type": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "map_style": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "thumbnail_url": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "collection_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "uploaded_at": {
                        "type": "string"
                    },
                    "source_updated_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "season_pivot": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "order_index": {
                                "type": "string"
                            },
                            "is_active": {
                                "type": "boolean"
                            }
                        },
                        "required": [
                            "id",
                            "order_index",
                            "is_active"
                        ]
                    }
                },
                "required": [
                    "id",
                    "uid",
                    "nadeo_map_id",
                    "name",
                    "author_account_id",
                    "author_name",
                    "author_time",
                    "gold_time",
                    "silver_time",
                    "bronze_time",
                    "map_type",
                    "map_style",
                    "thumbnail_url",
                    "collection_name",
                    "uploaded_at",
                    "source_updated_at",
                    "created_at",
                    "updated_at"
                ],
                "title": "MapResource"
            },
...
            "SeasonMapPlayerRecordResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "season_id": {
                        "type": "integer"
                    },
                    "map_id": {
                        "type": "integer"
                    },
                    "map": {
                        "$ref": "#/components/schemas/MapResource"
                    },
                    "trackmania_player_id": {
                        "type": "integer"
                    },
                    "player": {
                        "$ref": "#/components/schemas/TrackmaniaPlayerResource"
                    },
                    "global_position": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "current_position": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "time_ms": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "baseline_time_ms": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "first_seen_at": {
                        "type": "string"
                    },
                    "last_seen_at": {
                        "type": "string"
                    },
                    "last_improved_at": {
                        "type": "string"
                    },
                    "total_improvements": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "season_id",
                    "map_id",
                    "trackmania_player_id",
                    "global_position",
                    "current_position",
                    "time_ms",
                    "baseline_time_ms",
                    "first_seen_at",
                    "last_seen_at",
                    "last_improved_at",
                    "total_improvements",
                    "created_at",
                    "updated_at"
                ],
                "title": "SeasonMapPlayerRecordResource"
            },
            "SeasonResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "starts_at": {
                        "type": "string"
                    },
                    "ends_at": {
                        "type": "string"
                    },
                    "status": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "string",
                                "enum": [
                                    "draft"
                                ]
                            }
                        ]
                    },
                    "finalized_at": {
                        "type": "string"
                    },
                    "finalized_by_user_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "created_by_user_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "created_by": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "discord_username": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "name",
                            "discord_username"
                        ]
                    },
                    "maps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MapResource"
                        }
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "slug",
                    "description",
                    "starts_at",
                    "ends_at",
                    "status",
                    "finalized_at",
                    "finalized_by_user_id",
                    "created_by_user_id",
                    "created_at",
                    "updated_at"
                ],
                "title": "SeasonResource"
            },
...
            "TrackmaniaPlayerResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "account_id": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "zone_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "zone_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "last_seen_in_club_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "account_id",
                    "display_name",
                    "zone_id",
                    "zone_name",
                    "is_active",
                    "last_seen_in_club_at",
                    "created_at",
                    "updated_at"
                ],
                "title": "TrackmaniaPlayerResource"
            },
...
        },
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}